Re-duplicated `IOApp` to avoid bincompat issues
